% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\name{make_study}
\alias{make_study}
\title{Synthesize SDTM data for a clinical study}
\usage{
make_study(dm, treatment_table, sampling_scheme)
}
\arguments{
\item{dm}{The DM domain as data frame.}

\item{treatment_table}{The treatment/randomization table as data frame.}

\item{sampling_scheme}{The sampling scheme as data frame.}
}
\value{
A list.
}
\description{
Synthesize SDTM data for a clinical study
}
