% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\name{make_study_fe}
\alias{make_study_fe}
\title{Synthesize SDTM data for a food effect study}
\usage{
make_study_fe()
}
\value{
A list of SDTM domains.
}
\description{
Synthesize SDTM data for a food effect study
}
