% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\name{make_study_itz_rifa}
\alias{make_study_itz_rifa}
\title{Synthesize SDTM data for a DDI study with rifampin and itraconazole}
\usage{
make_study_itz_rifa()
}
\value{
A list of SDTM domains.
}
\description{
Synthesize SDTM data for a DDI study with rifampin and itraconazole
}
