% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\name{make_study_rba}
\alias{make_study_rba}
\title{Synthesize SDTM data for a relative BA study}
\usage{
make_study_rba()
}
\value{
A list of SDTM domains.
}
\description{
Synthesize SDTM data for a relative BA study
}
