% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\name{make_study_ri}
\alias{make_study_ri}
\title{Synthesize SDTM data for a renal impairment study}
\usage{
make_study_ri()
}
\value{
A list of SDTM domains.
}
\description{
Synthesize SDTM data for a renal impairment study
}
