% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{randomization_table}
\alias{randomization_table}
\title{Create randomization table}
\usage{
randomization_table(
  dm,
  EXDOSE = 100,
  EXDOSFRM = "TABLET",
  EXTRT = "RS2023",
  EXROUTE = "ORAL",
  FOOD = 0,
  sequence = data.frame(SEQUENCE = c("AB", "BA"), ARM = c("AB", "BA"), ARMCD = c("AB",
    "BA")),
  trtdy = c(1, 8),
  treatment = data.frame(TREATMENT = c("A", "B"), EXDOSE = c(500, 250))
)
}
\arguments{
\item{dm}{The DM domain as data frame}

\item{EXDOSE}{The dose as numeric.}

\item{EXDOSFRM}{The dosing form as character.}

\item{EXTRT}{The treatment as character.}

\item{EXROUTE}{The dosing route as character.}

\item{FOOD}{Fed administration as numeric.}

\item{sequence}{The sequences as vector of character strings where A, B, C,
etc. identify the treatment codes.}

\item{trtdy}{The treatment day(s)}

\item{treatment}{A data frame with TREATMENT giving the treatment codes and
further columns specifying the respective conditions in the treatment
periods.}
}
\value{
The randomization scheme as data frame.
}
\description{
Create randomization table
}
