% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{sad_table}
\alias{sad_table}
\title{Make a treatment table for SAD study}
\usage{
sad_table(
  dm,
  treatment = data.frame(EXDOSE = c(50, 100, 200, 300), N = c(3, 6, 3, 3))
)
}
\arguments{
\item{dm}{The DM domain as data frame.}

\item{treatment}{The treatments as data frame.}
}
\value{
A data frame.
}
\description{
Make a treatment table for SAD study
}
