% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdtm_data_ddi}
\alias{sdtm_data_ddi}
\title{SDTM data for a clinical DDI study with itraconazole and rifampicin}
\format{
A list of data frames corresponding to the SDTM domains DM, VS, LB, EX and PC.
}
\source{
generated using `sdtm_data_itz_rifa()'
}
\usage{
sdtm_data_ddi
}
\description{
Synthetic SDTM data for a fictional DDI study with the
fictional drug 'examplinib'.
}
\keyword{datasets}
