% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdtm_data_rba}
\alias{sdtm_data_rba}
\title{SDTM data for a relative bioavailabiliity study}
\format{
A list of data frames corresponding to the SDTM domains DM, VS, LB, EX and PC.
}
\source{
generated using `make_study_rba()'
}
\usage{
sdtm_data_rba
}
\description{
Synthetic SDTM data for a fictional relative bioavailability study with the
fictional drug 'examplinib'.
}
\keyword{datasets}
