% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdtm_data_ri}
\alias{sdtm_data_ri}
\title{SDTM data for a clinical study in patients with renal impairment}
\format{
A list of data frames corresponding to the SDTM domains DM, VS, LB, EX and PC.
}
\source{
generated using `sdtm_data_ri()'
}
\usage{
sdtm_data_ri
}
\description{
Synthetic SDTM data for a fictional DDI study with the
fictional drug 'examplinib'.
}
\keyword{datasets}
