% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdtm_data_sad}
\alias{sdtm_data_sad}
\title{SDTM data for a SAD study}
\format{
A list of data frames corresponding to the SDTM domains DM, VS, LB, EX and PC.
}
\source{
generated using `make_study_sad()'
}
\usage{
sdtm_data_sad
}
\description{
Synthetic SDTM data for a fictional single ascending dose study with the
fictional drug 'examplinib'.
}
\keyword{datasets}
