% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjects.R
\name{subject_baseline_data}
\alias{subject_baseline_data}
\title{Generate subject baseline data}
\usage{
subject_baseline_data(dm, vs, lb)
}
\arguments{
\item{dm}{The DM domain as data.frame.}

\item{vs}{The VS domain as data.frame.}

\item{lb}{The LB domain as data.frame.}
}
\value{
A data frame.
}
\description{
Generate subject baseline data
}
\keyword{internal}
