% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{synthesize_sd_ex}
\alias{synthesize_sd_ex}
\title{Create EX domain for single-dose studies}
\usage{
synthesize_sd_ex(
  dm,
  treatment_table,
  adminday = 1,
  drug = "RS2023",
  route = "ORAL",
  form = "TABLET",
  dose = 500
)
}
\arguments{
\item{dm}{The DM domain as data frame.}

\item{adminday}{The administration day(s) as numeric.}

\item{drug}{The drug as character.}

\item{route}{The administration route as character.}

\item{form}{The dosage form as character.}

\item{dose}{The dose as numeric.}

\item{randomization}{The randomization table as data frame.}
}
\value{
A data frame.
}
\description{
Create EX domain for single-dose studies
}
