% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{randomization_table}
\alias{randomization_table}
\title{Create randomization table}
\usage{
randomization_table(
  dm,
  sequence = c("AB", "BA"),
  adminday = c(1, 8),
  treatment = data.frame(TREATMENT = c("A", "B"), DOSE = c(500, 250))
)
}
\arguments{
\item{dm}{The DM domain as data frame}

\item{sequence}{The sequences as vector of character strings where A, B, C,
etc. identify the treatment codes.}

\item{adminday}{A vector of the treatment days as numeric.}

\item{treatment}{A data frame with TREATMENT giving the treatment codes and
further columns specifying the respective conditions in the treatment
periods.}
}
\value{
The randomization scheme as data frame.
}
\description{
Create randomization table
}
